Added documentation and moved reducers to the common pattern/reduce.hpp header
